define common sorter helper
