+++ app/models/queries/family_datatable_query.rb
@@ -1,57 +1,5 @@
 module Queries
   class FamilyDatatableQuery
-    # Helpers driving the sort query
-    module Sorter
-      # @method sort_query(query, order_by)
-      # Performs the sort operation on the query based on the order_by params.
-      # When the order_by params contain the 'name' or 'verification_due' column key, perform sort via an aggregation.
-      # Otherwise, perform the sort operation directly on the Mongoid query.
-      #
-      # @param [MongoidCriteria] query The query to be sorted.
-      # @param [Hash] order_by The sort column key and direction param hash.
-      #
-      # @return [Array<Family>|Mongoid::Criteria] The sorted array of Family objects if using an aggregated sort, or a Mongoid::Criteria object if using a direct sort.
-      def sort_query(query, order_by)
-        case order_by.keys.first
-        when 'name'
-          aggregate_sort(query, order_by) # perform the sort via an aggregation and return the sorted array of Family objects
-        else
-          query.order_by(order_by) # perform the sort directly on the Mongoid query and return the sorted Mongoid::Criteria object
-        end
-      end
-
-      # @method aggregate_sort(scope, order_by)
-      # Perform the sort operation on the query based on the order_by params via an aggregation.
-      #
-      # @param [MongoidCriteria] scope The query to be sorted.
-      # @param [Hash] order_by The sort column key and direction param hash.
-      #
-      # @return [Array<Family>|Mongoid::Criteria] The sorted array of Family objects if using an aggregated sort, or a Mongoid::Criteria object if using a direct sort.
-      def aggregate_sort(query, order_by)
-        pipeline = pipeline_for_sort_column(order_by)
-        # aggregate returns json, so we need to transform back to Family objects for the mongoid datatable to handle
-        ids = query.collection.aggregate(pipeline).map { |doc| doc["_id"] }
-        families = Family.where(:_id.in => ids).to_a
-        ids.map { |id| families.find { |family| family.id == id } }
-      end
-
-      # @method pipeline_for_sort_column(order_by)
-      # Construct the aggregation pipeline for sorting the query based on the order_by params.
-      #
-      # @param [Hash] order_by The sort column key and direction param hash.
-      #
-      # @return [Array] The pipeline array containing the aggregation stages for the order_by column key.
-      def pipeline_for_sort_column(order_by)
-        sort_column, sort_direction = order_by.to_a.flatten
-        sort_direction = sort_direction == :asc ? 1 : -1
-        base_pipeline = case sort_column
-                        when 'name'
-                          Family.sort_by_primary_full_name_pipeline(sort_direction)
-                        end
-        base_pipeline + [{:$skip => @skip}, {:$limit => @limit}]
-      end
-    end
-
     include Sorter
 
     attr_reader :search_string, :custom_attributes
@@ -65,27 +13,11 @@ module Queries
       @custom_attributes = attributes
       @skip = 0
       @limit = 25
+      @aggregatable_columns = {"name" => :sort_by_primary_full_name_pipeline}
     end
 
     def person_search search_string
       return Family if search_string.blank?
-    end
-
-   
-
-    def sort_by_name_col(query, sort_direction)
-      # Family name column is calculated using Family.family_members.primary_applicant.full_name
-      # primary_applicant is a Family model method, not a DB field
-      # full_name is a Person model method, not a DB field
-      # use an aggregation to perform the sort on name
-
-      # build the pipeline to sort by primary applicant's full name
-      pipeline = Family.sort_by_primary_full_name_pipeline(sort_direction)
-      pipeline += [{:$skip => @skip}, {:$limit => @limit}]
-      # aggregate returns json, so we need to transform back to Family objects for the mongoid datatable to handle
-      ids = query.collection.aggregate(pipeline).map { |doc| doc["_id"] }
-      families = Family.where(:_id.in => ids).to_a
-      ids.map { |id| families.find { |family| family.id == id } }
     end
 
     def build_scope
